Simplify testsuite driver
