+++ .github/workflows/ci.yml
@@ -43,7 +43,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: dtolnay/rust-toolchain@stable
       - uses: Swatinem/rust-cache@v2
-      - run: cargo run -p test-runner -- run
+      - run: cargo run -p test-runner
 
   fmt:
     name: Rustfmt
